# docs/swagger.yaml
definitions:
  ncobase_common_resp.Exception:
    properties:
      code:
        description: Business code
        type: integer
      data:
        description: Response data
      errors:
        description: Validation errors
      message:
        description: Message
        type: string
      status:
        description: HTTP status
        type: integer
    type: object
  ncobase_common_types.JSON:
    additionalProperties: {}
    type: object
  structs.Captcha:
    properties:
      id:
        type: string
      solution:
        type: string
    required:
    - id
    - solution
    type: object
  structs.CasbinRuleBody:
    properties:
      p_type:
        type: string
      v0:
        type: string
      v1:
        type: string
      v2:
        type: string
      v3:
        type: string
      v4:
        type: string
      v5:
        type: string
    required:
    - p_type
    - v0
    - v1
    - v2
    type: object
...
  structs.LoginBody:
    properties:
      captcha:
        $ref: '#/definitions/structs.Captcha'
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  structs.RegisterBody:
    properties:
      display_name:
        type: string
      phone:
        type: string
      register_token:
        type: string
      short_bio:
        type: string
      tenant:
        type: string
      username:
        type: string
    required:
    - display_name
    - register_token
    - username
    type: object
  structs.SendCodeBody:
    properties:
      email:
        type: string
      phone:
        type: string
    type: object
...
info:
  contact: {}
  description: Next-Gen Modular Framework
  title: Ncobase
  version: 0.1.0
paths:
...
  /v1/authorize/{code}:
    get:
      consumes:
      - application/json
      description: Verify the provided code.
      parameters:
      - description: Verification code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/ncobase_common_types.JSON'
            - properties:
                access_token:
                  type: string
                email:
                  type: string
                id:
                  type: string
                register_token:
                  type: string
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      summary: Verify code
      tags:
      - authorization
      - authentication
  /v1/authorize/send:
    post:
      consumes:
      - application/json
      description: Send a verification code to the specified destination.
      parameters:
      - description: SendCodeBody object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/structs.SendCodeBody'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/ncobase_common_types.JSON'
            - properties:
                registered:
                  type: boolean
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      summary: Send verification code
      tags:
      - authorization
  /v1/captcha/{captcha_id}:
    get:
      description: Stream a captcha image.
      parameters:
      - description: Captcha ID With Extension (png, wav)
        in: path
        name: captcha_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      summary: Stream captcha
      tags:
      - authentication
  /v1/captcha/generate:
    get:
      description: Generate a captcha image.
      parameters:
      - description: Captcha type
        enum:
        - png
        - wav
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/ncobase_common_types.JSON'
            - properties:
                id:
                  type: string
                url:
                  type: string
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      summary: Generate captcha
      tags:
      - authentication
  /v1/captcha/validate:
    post:
      consumes:
      - application/json
      description: Validate a captcha code.
      parameters:
      - description: Captcha object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/structs.Captcha'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/ncobase_common_types.JSON'
            - properties:
                message:
                  type: string
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      summary: Validate captcha
      tags:
      - authentication
  /v1/login:
    post:
      consumes:
      - application/json
      description: Log in a user.
      parameters:
      - description: LoginBody object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/structs.LoginBody'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/ncobase_common_types.JSON'
            - properties:
                access_token:
                  type: string
                id:
                  type: string
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      summary: Login
      tags:
      - authentication
  /v1/logout:
    post:
      description: Logout the current user.
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Logout
      tags:
      - authentication
...
  /v1/pols:
    get:
      description: Retrieve a list of Casbin rules.
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: offset
        type: integer
      - in: query
        name: p_type
        type: string
      - in: query
        name: v0
        type: string
      - in: query
        name: v1
        type: string
      - in: query
        name: v2
        type: string
      - in: query
        name: v3
        type: string
      - in: query
        name: v4
        type: string
      - in: query
        name: v5
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            items:
              $ref: '#/definitions/structs.CasbinRuleBody'
            type: array
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: List Casbin rules
      tags:
      - casbin
    post:
      consumes:
      - application/json
      description: Create a new Casbin rule.
      parameters:
      - description: CasbinRuleBody object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/structs.CasbinRuleBody'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/structs.CasbinRuleBody'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Create Casbin rule
      tags:
      - casbin
  /v1/pols/{id}:
    delete:
      description: Delete an existing Casbin rule.
      parameters:
      - description: Casbin rule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Delete Casbin rule
      tags:
      - casbin
    get:
      description: Retrieve details of a Casbin rule.
      parameters:
      - description: Casbin rule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/structs.CasbinRuleBody'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Get Casbin rule
      tags:
      - casbin
    put:
      consumes:
      - application/json
      description: Update an existing Casbin rule, either fully or partially.
      parameters:
      - description: Casbin rule ID
        in: path
        name: id
        required: true
        type: string
      - description: CasbinRuleBody object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/structs.CasbinRuleBody'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/structs.CasbinRuleBody'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      security:
      - Bearer: []
      summary: Update Casbin rule
      tags:
      - casbin
  /v1/register:
    post:
      consumes:
      - application/json
      description: Register a new user.
      parameters:
      - description: RegisterBody object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/structs.RegisterBody'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/ncobase_common_types.JSON'
            - properties:
                access_token:
                  type: string
                id:
                  type: string
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/ncobase_common_resp.Exception'
      summary: Register
      tags:
      - authentication
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"